// src/shared/api/generated/types/AdminNewsCategoryDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type AdminNewsCategoryDto = {
  /**
   * @type number
   */
  id: number;
  /**
   * @type string
   */
  slug: string;
  /**
   * @type string
   */
  title: string;
  /**
   * @type string
   */
  description?: string | null;
  /**
   * @type number
   */
  order: number;
  /**
   * @type boolean
   */
  isActive: boolean;
  /**
   * @type number
   */
  newsCount: number;
  /**
   * @type string, date-time
   */
  createdAt: string;
  /**
   * @type string, date-time
   */
  updatedAt: string;
};